Fix PHP 8.4 compatibility for unit testsuite
- Add return types, strict typing casts, and null-safety fixes across core
- Fix bootstrap-assertions re-exec for isolated PHPUnit processes
- Update unit tests for Symfony 7, bool image toolkit returns, and markup APIs
- Sync scaffold ht.router.php and correct SearchPageRepository default logic

+++ core/lib/Drupal/Core/Batch/Percentage.php
@@ -37,7 +37,7 @@ class Percentage
         } else {
             // We add a new digit at 200, 2000, etc. (since, for example, 199/200
             // would round up to 100% if we didn't).
-            $decimal_places = max(0, floor(log10($total / 2.0)) - 1);
+            $decimal_places = max(0, (int) floor(log10($total / 2.0)) - 1);
             do {
                 // Calculate the percentage to the specified number of decimal places.
                 $percentage = sprintf('%01.' . $decimal_places . 'f', round($current / $total * 100, $decimal_places));
